# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open Local — open source marketplace for local vendors and producers
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Vendor {
  id: number;
  name: string;
  slug: string;
  tagline: string;
  description: string;
  category: string;
  location: string;
  region: string;
  contactEmail: string;
  /** @nullable */
  websiteUrl: string | null;
  imageUrl: string;
  established: number;
  featured: boolean;
  createdAt: string;
}

export interface VendorInput {
  name: string;
  slug: string;
  tagline: string;
  description: string;
  category: string;
  location: string;
  region: string;
  contactEmail: string;
  /** @nullable */
  websiteUrl?: string | null;
  imageUrl: string;
  established: number;
  featured?: boolean;
}

export interface VendorUpdate {
  name?: string;
  slug?: string;
  tagline?: string;
  description?: string;
  category?: string;
  location?: string;
  region?: string;
  contactEmail?: string;
  /** @nullable */
  websiteUrl?: string | null;
  imageUrl?: string;
  established?: number;
  featured?: boolean;
}

export interface Product {
  id: number;
  vendorId: number;
  name: string;
  description: string;
  priceCents: number;
  unit: string;
  category: string;
  imageUrl: string;
  inStock: boolean;
  featured: boolean;
  createdAt: string;
}

export interface ProductWithVendor {
  id: number;
  vendorId: number;
  name: string;
  description: string;
  priceCents: number;
  unit: string;
  category: string;
  imageUrl: string;
  inStock: boolean;
  featured: boolean;
  createdAt: string;
  vendorName: string;
  vendorSlug: string;
  vendorLocation: string;
  vendorCategory: string;
}

export interface ProductInput {
  vendorId: number;
  name: string;
  description: string;
  priceCents: number;
  unit: string;
  category: string;
  imageUrl: string;
  inStock?: boolean;
  featured?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  priceCents?: number;
  unit?: string;
  category?: string;
  imageUrl?: string;
  inStock?: boolean;
  featured?: boolean;
}

export interface MarketplaceStats {
  vendorCount: number;
  productCount: number;
  locationCount: number;
  categoryCount: number;
  inStockCount: number;
  featuredVendorCount: number;
}

export interface LocationCount {
  location: string;
  region: string;
  vendorCount: number;
}

export interface CategoryCount {
  name: string;
  count: number;
}

export interface CategoryBreakdown {
  vendorCategories: CategoryCount[];
  productCategories: CategoryCount[];
}

export type ListVendorsParams = {
  search?: string;
  category?: string;
  location?: string;
  featured?: boolean;
};

export type ListProductsParams = {
  search?: string;
  category?: string;
  vendorId?: number;
  featured?: boolean;
  inStock?: boolean;
};