# lib/api-zod/src/generated/types/locationCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open Local — open source marketplace for local vendors and producers
 * OpenAPI spec version: 0.1.0
 */

export interface LocationCount {
  location: string;
  region: string;
  vendorCount: number;
}