# lib/api-zod/src/generated/types/productWithVendor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open Local — open source marketplace for local vendors and producers
 * OpenAPI spec version: 0.1.0
 */

export interface ProductWithVendor {
  id: number;
  vendorId: number;
  name: string;
  description: string;
  priceCents: number;
  unit: string;
  category: string;
  imageUrl: string;
  inStock: boolean;
  featured: boolean;
  createdAt: Date;
  vendorName: string;
  vendorSlug: string;
  vendorLocation: string;
  vendorCategory: string;
}