# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open Local — open source marketplace for local vendors and producers
 * OpenAPI spec version: 0.1.0
 */

export interface Vendor {
  id: number;
  name: string;
  slug: string;
  tagline: string;
  description: string;
  category: string;
  location: string;
  region: string;
  contactEmail: string;
  /** @nullable */
  websiteUrl: string | null;
  imageUrl: string;
  established: number;
  featured: boolean;
  createdAt: Date;
}